Ensure AIO container_tech/install_method vars are namespaced
To ensure there are no clashes with the inventory and other uses
beyond the AIO bootstrap, we namespace the container_tech variable
and the install_method variable. This also allows us to use those
variables as the common base in all conditionals, rather than repeat
the use of the env var lookup here and there.

We also tidy up a bit of content in the user_variables.yml template and
the bootstrap-aio.yml playbook.

Finally, we remove the unnecessary specification of the
'bootstrap_host_scenario', 'install_method' and 'openstack_confd_entries'
variables in the playbook, given that they're now in the role defaults.

Change-Id: I5c276a2efed5e1b300fc72f2767abb008199571c