Implement AIO changes to allow test usage by roles
In this patch we change the 'scenario' for the AIO bootstrap
to allow it to be overloaded with any additional service names
which correspond to an appropriate conf.d AIO configuration file.

This allows us to implement jobs using a scenario called something
like 'aio_lxc_octavia' and it will implement the base AIO and add
the octavia service.

In order to make use of this for the experimental jobs, we change
the job names so that the appropriate services are deployed.

Related-bug: #1759784
Change-Id: If6b7e401021ea57a908a616ec8b86e7aa6798558
(cherry picked from commit ada265c8449334538a011300785634774787d54d)