Ensure that a consistent mirror is used for RDO
This patch replicates the code added to openstack-ansible-tests into
the integrated repo. It obtains a URL for the RDO repo at the start
of a job which then has consistent contents for the duration of the job.

Change-Id: I6492c66ccc9d2b5b6de9ab6d08b377ac9e18e153
(cherry picked from commit 11a71e11313f7a424558c7d78f1a00e3b58b9565)