Set SUSE ceph-distro job to non-voting
In https://review.openstack.org/616481 we switched OSA stable/rocky
from ceph-ansible 3.2 to 3.1, but https://review.openstack.org/614463
made the job voting before that merged. The 3.1 builds for SUSE do not
work (see related bug), so we switch this job to non-voting. Whenever
the issue is fixed upstream, we can revert this (and add the missing
gate job too).

Change-Id: I70059d1b0f9eb774519420b14239ab7e3081e7eb
Related-Bug: #1805637