Document how to use the utility container for service setup
If the service setup host has been set to the utility container, the python
interpreter needs to be set to the venv rather than the system otherwise the
shade/openstacksdk library will not be found.

In this patch we add a commented set of overrides into user_variables.yml
as a convenience for users as this will be a relatively common requirement.

Depends-On: https://review.openstack.org/621189
Depends-On: https://review.openstack.org/621353
Depends-On: https://review.openstack.org/621355
Depends-On: https://review.openstack.org/621655
Depends-On: https://review.openstack.org/621354
Depends-On: https://review.openstack.org/622523
Change-Id: I16f5145f5d64ebe757ede8b4efeda2ffea5957ca
(cherry picked from commit e29bc2203fbf0c466f0f2eeaabe04653364007d9)