Document how to use the utility container for service setup
If the service setup host has been set to the utility container, the python
interpreter needs to be set to the venv rather than the system otherwise the
shade/openstacksdk library will not be found.

In this patch we add a commented set of overrides into user_variables.yml
as a convenience for users as this will be a relatively common requirement.

Depends-On: https://review.openstack.org/620644
Depends-On: https://review.openstack.org/621173
Depends-On: https://review.openstack.org/621171
Depends-On: https://review.openstack.org/621175
Depends-On: https://review.openstack.org/621202
Depends-On: https://review.openstack.org/621178
Change-Id: I16f5145f5d64ebe757ede8b4efeda2ffea5957ca