bootstrap: stop creating swap
We used to create swap because we couldn't get a deployment to
fit in our upstream CI systems.  The average swap usage now is
1-2 MB but it can take up to 15-20 minutes to create the swap
which can cause jobs to timeout.

This patch stops us from creating swap as it's not necessary,
if we run out of memory, we should probably look into that issue
as an AIO inside 8GB of memory is plenty!

Change-Id: I587d1fe984e8b9040e341b69aeb450ec5459d60f