Zuul: Simplify the integrated test playbooks
When using the integrated repo for testing, only the openstack-ansible
repository can be used to kick off the tests - so if any other repo
uses it, it will always be the same path.

The existing implementation modeled from the openstack-ansible-tests
repo which implemented a job and set of playbooks which would allow
the tests to execute from any one of the other repositories - which
is not the case here.

There is therefore no reason to have this path be dynamic in any way,
so we hard code it to where we know zuul has it.

We also remove other unnecessary fact setting and ensure that bash is
used when running gate-check-commit.

Change-Id: I17780e48cac97b1f3888b8cb264c284d8ece44a1