Ensure that control plane hosts are also updated correctly
In queens the default for neutron_agents was set to deploy on metal.
so the exclusion of neutron_agents in the upgrade process for the
setup-hosts playbook end up excluding the whole control plane, which
is not good. One thing that happens for the Q->R upgrade is that the
apt proxy configuration is not removed, then all apt actions fail.

To ensure that we update the control plane as expected, we remove
the neutron_agents group from the exclusion.

Change-Id: Ic9453b042ac866abe55723dd16725f58e9453718
(cherry picked from commit 38b61e1027b3440f408c36f24fdcbe0037283c4d)