Add automated migration of neutron agents to bare metal
In Id16b1bcd67bf075a8677206a205ece010e7b014c the environment files
were changed in order to ensure that new deployments have neutron
agents on the bare metal host, not in containers. However, after
upgrading from P to Q the agents are still in a container and it
was left to the deployer to figure out how to do the migration.

With thanks to the co-authors this process has now been worked out
and this patch adds automation for that process to make it simpler
for deployers.

Note that there is no rebalancing implemented. It seems that if the
default deployment settings are used, neutron handles the rebalancing
on its own.

Also note that no neutron lbaasv2 changes are included - this can be
implemented in a follow up patch if need be. The related bug below
provides the requisite database change, but I do not have a test
environment which can verify any other steps that may be required.

Change-Id: I88836acef3d4033f9b887093f44e7089c92bbe6f
Co-Authored-By: Daniel Marks <daniel.marks@codecentric.de>
Co-Authored-By: Christian Zunker <christian.zunker@codecentric.de>
Co-Authored-By: James Denton <james.denton@rackspace.com>
Closes-Bug: #1785592
Related-Bug: #1804770