Add http proxy test scenario
This patch adds a test scenario which deploys a squid proxy on
the AIO host and uses deployment_environment_variables to direct
all http and https traffic through that proxy.

In order to make the test reasonably realistic eth0 is removed
from all containers which means they have no default route
and must retrieve all external dependencies via the proxy.

As eth0 is removed we can no longer use it to NAT the traffic to
the neutron public ip address range during tempest tests. Instead
we add static routes to those address ranges to eth1 in the
containers and expect the host to forward via br-mgmt.

Depends-On: https://review.openstack.org/625312
Depends-On: https://review.openstack.org/625670

Change-Id: I554cfd149374f0863ac058f26862b5439c5d53d8