Set DynamicUser in systemd-journal-upload unit file
At least on Ubuntu 18.04, the `systemd-journal-remote` doesn't create
the `systemd-journal-upload` user, since the service is supposed to run
with `DynamicUser=True`.

This commit ensures that `DynamicUser` and `StateDirectory` are set
appropriately on all Systemd versions that support them (i.e. >= 235).

It requires https://review.openstack.org/#/c/630664/ to be merged first.

Change-Id: I743e936df9e69ef5532b7ec69e29057d4f1f8ef2
Closes-Bug: 1805847