Fix ceph-ansible role-requirements
Ceph has removed the individual repos of common, osds, mons, defaults...
this uses the new combined ceph-ansible repo.

Using the combined ceph-ansible repo breaks ansible-galaxy (because
ceph-ansible is not a role), so we also need to switch the default for
the ansible bootstrap to use git instead, and we need to adjust the lint
test to use git instead of ansible-galaxy.

Change-Id: I73141e9b4ef4cfa83b9ad3c434be91a270ecc99a