Define OSA clone dir in the openstack-ansible.sh script
When the openstack-ansible.sh was extracted from bootstrap.sh,
the resolving from the OSA_CLONE_DIR was broken.

Due to this, the OSA_CLONE_DIR value is empty, and the comparison
in the openstack-ansible is always true, which is not what we
want for external use.

Change-Id: I6a05d3b73838fa646a635831d655f7470df3378f
(cherry picked from commit ecdd3822b6007eb804d9e3681959cc382859103f)