nspawn: fix restart-service.yml service list filtering
The service list filtering is supposed to look for all services
that are prefixed with 'service_name', however, we are not doing
that in the regex.

When using nspawn and restarting a service such as 'nova', it matches
the systemd-nspawn service for the container and attempts to reload
it which fails.

This fixes it by forcing it to check for services *prefixed* with
that variable.

Change-Id: Ib1a3975f2722127eada2dd957300fd306abbe6d9