Add option to disable container builds on specific hosts
This change adds an option to disable any and all containers within a given host
The new option "no_containers" is a boolean option and has a default value of
false. When set true no containers will be created in inventory for a given machine.

This will allow deployers to simply disable container builds on specific machines
when nessisary. This is usefule when a deployer simply wants to deploy openstack
in a flattened environment. This has also been useful in cases where
infrastructure hosts may reside within virtualization solutions like OpenStack,
VirtualBox, or VMWare which implement strong port security rules which makes
container networking difficult or impossible.

To implement the change simply add the "no_containers: true|false" option to
a host entry in the openstack_user_config.yml

Change-Id: I8bed3e498a431e2683956e1009d9d9ece9fdb272
Signed-off-by: Kevin Carter <kevin@cloudnull.com>