Update the heat service egg name to openstack_heat
In https://review.openstack.org/606160 the dist name was changed
from 'heat' to 'openstack-heat'. The egg name needs to be updated
or the package will fail to install or be usable.

This is needed in Rocky because the change for the heat package name
was backported. As of the recent Rocky SHA bump[1], Heat will no
longer build in OSA without this fix.

[1] https://opendev.org/openstack/openstack-ansible/commit/334096f6a9e84acbceca64ef08d1390fdb790da3

Related: https://review.opendev.org/#/c/650438/
Change-Id: Ida55bc030800c512d64bb715f704697875cd5738
(cherry picked from commit 3621ce76b706a1707cc986859f4c7b5eaa0b1f23)