Use runtime venv to determine OSA version
With the introduction of c44fbcbbe3b6d88e9393503e92ac670f9e811148,
setuptools is used to determine the current OSA version. However,
currently it relies on setuptools being present on the host. This
is an unnecessary dependency given that we have setuptools in the
runtime venv.

Rather than add more dependencies on the host, we now use the
runtime venv to check for the version of OSA.

Change-Id: I5bf11906c87d308567af483263e753ab4f9e2815