Convert include to either import_playbook or include_tasks
This change resolves the deprecation warnings for the use of "include"
which has been removed upstream. Playbooks will now use "import_playbook"
and tasks will now use "include_tasks". While this change attempts to
resolve all of the include issues several playbooks will need to be
refactored to resolve the use of include with variables as this is not
a pattern that is supported by mainline ansible.

Change-Id: I8fdb2f9f75f38986ba1dc9f93e274749c49e5c67
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>