Remove broken conditions when including playbooks
The conditions in our setup playbooks have been broken for a long time
however they've not caused any issues because they're not deprecated
because the use of "When" with playbook includes is simply ignored.
This change removes the conditions which cleans up our setup playbooks.
The conditions to exit or execute a given playbook has been moved into
the playbook. If the condition is not met then the meta tag is used
to exit the playbook and continue.

Change-Id: I4e1c2eb1ebb0dfb01e5c98e0fd13272943e9a97c
Signed-off-by: Kevin Carter <kevin@cloudnull.com>