nova: restart instead of reloading services
There is a bug inside oslo.service which causes calls a stop,
reset and start instead of reset only on SIGHUP which causes
a very inconsistent state of services.

Until this is resolved, we should not reload services but restart
them only.  This patch should be reverted once the related bug
is fixed.

This patch also removes SUSE jobs from check/gate and into the
experimental queue as they do not have full maintainers at the
moment.

In addition, it moves jobs that have been failing for a long time
such as the aio_distro_lxc to experimental as well, as they have
been problematic.

Change-Id: I61e340a4ef5f0980e8c36611c0f78cd06946636b
Related-Bug: #1715374