Always set pip vars for tempest install
In stable/rocky, there is no distro install method for os_tempest
so it will always install from source. However, in distro installs
there is no repo container and no constraints are set, so the
tempest installation is currently unconstrained which delivers
inconsistent results from one day to the next.

To ensure that the installation is constrained properly, we make
sure that the pip vars are set and the correct vars are used when
installing tempest.

Given that the tempest_ vars in distro_install are not used, we
remove them from the file to reduce confusion.

Change-Id: I851e50902d3866f8f28ebc6f172bc61d0a4778c3