Pass ceph_mons as a list of inventory hosts
Since change Idb9a178ca2165920804eee12a54419722aa64f51, we no longer
attempt to directly SSH to the ceph mons. Instead, we use the ansible
wait_for_connection module. One advantage to using this module is we
can pass it an ansible inventory host instead of an IP address, and
it will use the configured connection settings for that inventory
host, such as the container connection plugin.

This allows the ceph-client tasks requiring mon access to complete
without a hard dependency on an SSH connection to the mon. With this
change, we will simply pass the ceph-mon inventory hostnames to the
ceph_client role rather than a list of IP addresses.

Change-Id: Ia94ba882c2a8f568a2dcca92ee381f5833bd50af