barbican: run without SSL inside gate for metal jobs
Metal jobs currently don't have HAproxy so we don't have any SSL
happening on the public network.  All tempest plugins have settings
to allow to test against the internal URL which avoids hitting the
SSL-less non-functional endpoint.  Barbican does not, so let's
just set it up with HTTP in the catalog for now so we can get it's
test to pass correctly.

Depends-On: https://review.openstack.org/648551
Change-Id: I5d22c24c331251779c696149c7c564c3baeed0d3