nova: refactor common playbook to use role instead of include_role
Since introducing distro. installs, we have two different role
inclusions, one for distro and one for source.  It seems that this
was done to send a different set of variables to each run.

With the recent build changes, all the branch variables are already
included using from openstack_services.yml so there's no need to
redefine them.

This means we can bring the two includes back to a normal role
reference, which resolves a few issues around using import_role and
brings back the ability to use tags only with this playbook.

Change-Id: I6c6651ce923530b1d2395cd387faf6e74c9b17d8