octavia: use br-lbaas IP for health managers
At the moment, there is a bug stopping us from being able to
properly supporting metal in Octavia.  In order to get the metal
gate to pass, we work around it by overriding the variable, however
we'll need to re-evaluate on how to be able to retrieve IPs of other
hosts scoped within a network on metal eventually to resovle this.

Change-Id: I5b9d4f386ac19be317602b30b08b8145f980de31