glance: stop using healthcheck endpoint
The healthcheck endpoint currently is not configured by us, which
means that it's pretty much a no-op right now.

It also causes issues with IPv6 addresses, where it fails to parse
the IP correctly and raises an exception.

    ValueError: invalid literal for int() with
    base 10: 'xxxx:x:x::xx]:9292'

This patch goes back to monitoring the root of the endpoint as that
probably delivers just as accurate of information in the current
state that we're in.

Change-Id: I859795307b88508dc1b894b39526f4289b33c1af