config: Enable OSA_CONFIG_DIR to customize configuration folder
We have some basics wired up in order to use a different config.
directory for running OpenStack Ansible.  However, a lot of those
are implemented as configuration argument to the dynamic inventory
script but that's impossible to pass normally when running Ansible.

This patch creates a new environment variable called OSA_CONFIG_DIR
which allows you to customize the directoy of your openstack_deploy
folder which means that you can use a deploy host for several diff.
environemnts.

Change-Id: Ia528ff449330fbb04be6cd0d03353fa3158e6694