nova: reduce playbooks runs to control/compute
During deployment or upgrades, there isn't a reason to deploy a
service after another.  There is no documented requirement for the
order of upgrades either.

We also start services in a specific order using the systemd_service
role which means that we can simplify this into two main role includes,
control plane and data plane (computes).  This is to make sure that
our upgrades are completed cleanly.

Depends-On: I506ba4fc6b134ef0b389a2647fe22ae6b1a17ace
Change-Id: I98ca698478e24b65b52b2b3b10052532b8425976