Ensure that pip/setuptools/wheel python wheels are built
All virtualenvs make use of pip/setuptools/wheel, but none of
them specify that these wheels need to be built and therefore
every venv build process reaches out to pypi to fetch them.

To ensure that the repo container has these built as wheels,
we set the utility container to build them so that these
packages will be available for other venvs to use.

Change-Id: I60f851e1ffa401c92221de14651502306e9b0581