Set CentOS 7 distro job to non-voting
We need to merge https://review.opendev.org/#/c/656891/ to unblock
the gate, but RDO packging for nova is broken[1] and is preventing
us from merging anything currently. We should make it non-voting to
get the gates working again, work on removing the nova-placement-api
package from the nova role, and then re-enable voting.

[1] https://review.opendev.org/#/c/657398/

Change-Id: Id9bd94923e05b3911a9dfd4e10c43f4c0ab81b9b