Remove credential sharing with nova
Various authtoken parameters are shifted up from the role defaults of
ironic, neutron, and placement so that nova can authenticate as these
services rather than itself when connecting to them as an API client.

Nova's configuration has been modififed not to require credential
sharing, so it will authenticate as itself to reach other APIs as
a client.

Depends-On: I021f85fd4d08e49876377f87a5cd2fddf54a236f
Change-Id: Ibc68c6b89821e6e4905d16a9115a9e42284cb0ec