Make scenario construction more generic
The bootstrap host scenario was very quirky because we had to a maintain
a list of "special" scenario keywords that are not actual services,
otherwise the role would fail when deploying non-existent AIO conf.d
files.

Instead, we can simply pass a full expanded scenario list (including
the "special" keywords and all service keywords and only deploy the
conf.d files that exist in the repo.

This will make scenarios much more straight forward to use and less
flaky when adding scenario variants that may not have a conf.d file
to deploy.

Change-Id: I621660f39c4e51db4b8063feee782aae5cd16840