Move data device detection to Ansible
Move the detection of the secondary data disk device to the
bootstrap-host role. This is important because the detection was
previously done in bash in the gate-check-commit, but since
bootstrap-aio is executed independently of gate-check-commit in the
gate, the data disk is no longer being used correctly. Now the
data disk will bootstrap reliably regardless of how bootstrap-aio
is run.

Change-Id: I2d49b6802d8cf99f6d7ac74ae4ce0ab8f6d1c7a5
(cherry picked from commit 4bc3e429c578ca2e77f79eecc6467725d9b0e6c9)