Document bootstrap host failure on leap 15
Bootstrap host is relying on hostname ansible module to set
hostname. On Leap 15, we are using /etc/os-release instead of
suse-release, which breaks the platform stdlib on python2 and
python3. There is no intention upstream to patch `platform`,
as this has been deprecated, to accept any workaround SUSE needs
to implement. Instead, it is advised to move to the `distro`
library. Hopefully, Ansible is already moving to this. Because
it's very internal to ansible, ansible vendors-in its own version
of it. When we'll upgrade ansible to 2.8 or above, we should get
an improvement (to be tested), or be able to fix it in Ansible
and in distro.

Without this patch, we don't know how to fix, nor who is
responsible of the fix. This should clarify things.

Change-Id: I4394755fbcc91c56e11037729a7683ff658a9f59