Update Ansible to 2.8.2, Ceph-Ansible to 4.0, Ceph to Nautilus
This patch updates to the current version of Ansible in the 2.8 series,
2.8.2.

Ceph-Ansible must be moved to version 4.0 to be compatible with Ansible
2.8 [1]. Older versions of Ceph-Ansible use syntax which is invalid for
Ansible 2.8.

In turn, Ceph-Ansible 4.0 can only install the Nautilus release of Ceph,
there is no overlap with the previous Ceph release as there has been
in the prior Ceph-Ansible 3.x series.

This patch updates provisioning of the OSA AIO ceph build to be
compatible with the ceph-volume tool [2], specifically ceph-volume
cannot prepare OSDs from loopback devices. The AIO boostrap now creates
LVM on top of the loopbacks which are then consumed by ceph-volume.

The configuration of radosgw is updated to account for the addition of
instance names to rados gateway configurations [3][4].

[1] http://docs.ceph.com/ceph-ansible/master/#releases
[2] http://docs.ceph.com/docs/nautilus/ceph-volume
[3] https://github.com/ceph/ceph-ansible/blob/367dce2894be37458b5129f4759633e0929f2b8b/roles/ceph-config/templates/ceph.conf.j2#L115-L116
[4] https://github.com/ceph/ceph-ansible/blob/367dce2894be37458b5129f4759633e0929f2b8b/roles/ceph-facts/tasks/facts.yml#L276-L280

Change-Id: I7075be2b06e81e1008bd393cf0fc5619dad24e7e