Bump etcd role to v3 capable SHA
etcd v3 is required by the upgraded version of Calico that is
compatiable with openstack versions Queens and later.

The systemd_service role is added with a null trackbranch and a SHA
pinned at the current HEAD of the role's stable/rocky branch. That
is necessary because the etcd role's tags which install etcd3 utilize
the systemd_service role to build a unit file for starting etcd.

Depends-On: I746b9dbb090dd484a36e826f06beda1eba3f55f4
Change-Id: I04bb2773d5e55e73f29b2c24013005ea814d3992
(cherry picked from commit 00a451c874aa63047ad0da4f34f62cf318d2ac78)