Fix mistral metal deploy
In AIO CI deploy, we are relying on aio_metal.yml.example to
determine what to set in the runtime env.d (it is copied there,
verbatim).

Sadly we added in env.d the support for mistral, without editing
this file. So, by default, the integrated will use lxc, as the
group for mistral will not have is_metal: true. Therefore
the lxc_hosts group will also be populated, and the deployment
is then using lxc for just this container. It is not respecting
the requirement for running 'on metal'.

This should fix it.

Change-Id: Icf29be45ceba24b31ff471157b630e23661709a5