Improve repo server haproxy healthcheck to verify repo synchronisation
This patch updates the healthcheck to test for the presence of the file
/var/www/repo/repo_sync_complete.
This addresses an issue where haproxy would direct traffic to redeployed
repo servers which were correctly serving the web root but had no
meaningful content.
The repo_sync_complete file is created by related changes to the
repo_server ansible role.

Partial-Bug: 1829930
Depends-On: https://review.opendev.org/660788/

Change-Id: Ib7b64b09413fbdd8c630311d437539c9f876fac0