Utility container venv is missing python libs for ansible openstack modules
Some service set up tasks use ansible modules that require external
python libraries. When using a utility container as a service setup host
these ansible tasks will fail if the required libraries are not present
in the utility venv.

Added openstacksdk to the list of pip packages installed to the container
as required for ansible 2.7 and above.

Change-Id: I25a89ed706cad0ae3376c466807f3ba024941c2a
Closes-Bug: #1830028