Load extra variables correctly in wrapper
The extra variables are meant to be loaded contextually if ansible
or ansible-playbook are executed in the OSA playbooks dir, or if the
openstack-ansible wrapper is explicitly called.

This is occurring correctly for openstack-ansible and ansible-playbook,
but extra vars are not being loaded correctly (despite being output
to console as if they're being loaded) when ad-hoc ansible commands
are run in the OSA context.

Change-Id: I023871df159413f9313d7a85820a50872da2b36b