Add aio_metal and aio_distro_metal openSUSE 15 jobs
opensuse-15 is a nodepool image set referring to the latest
enabled openSUSE 15.x release, currently pointing to openSUSE 15.1.

We'd like to transition away from 15.0 as soon as possible. As
part of the rename to 15 we can actually now test against the
latest minor leap 15.x release without continuous job maintenance
effort.

In order to find problems and test stability of the opensuse-15
distro_metal job we add a periodic job to monitor it for the next ~ 1
month and then redecide.

Change-Id: Ided0d56d2bd491142e9c098252ed5c0f8d04f831