Differentiate AIO /etc/hosts entries for localhost and aio1
This patch ensures that localhost resolves to a loopback address, and
aio1 resolves to the address assigned to br-mgmt.

Without this change rabbitmq is unable to function properly when it is
bound to the mgmt ip rather than 0.0.0.0 as there is a mismatch between
the hostname 'aio1' and the /etc/hosts entry for that ip.

The rabbitmq role corrects /etc/hosts entries for 127.0.0.1, but not
127.0.1.1 so the extra entry is removed. The aio1 hosts entry is changed
to resolve to 172.29.236.100.

Change-Id: I194fcab8b84723cff76ec09315e264ec7840ea64