Ceph-RGW HAProxy config skipped when shouldnt
Now that group ceph-rgw is in the default inventory (commit c700fdba)
the logic for haproxy_backend_nodes would always select the group
'ceph-rgw' as it will always be defined, regardless of length.

If using the ceph_rgws variable instead of the ceph-rgws group it
meant that the haproxy_backend_nodes equated to an empty group.
So the 'when' criteria in haproxy_service_config.yml did not get met
and the services config was not provisioned.

This fix checks the length of the ceph-rgws group.
If its not greater than zero, haproxy_backend_nodes is defaulted
back to ceph_rgws.

Change-Id: I7ac3e3ba9e64b38ec22704394fcd5f402dd9d7f8