Add unified linters check
Since now linters check are provided by openstack-ansible-tests repo
which we're trying to get rid of, this patch implements new job template
openstack-ansible-linters-jobs which aims to cover linters checks, that
will be runned by gate-check-commit script. It provides next checks:
- ansible-lint
- ansible syntax-check
- bashate
- flake8

Role for the check will be selected based on the zuul.project.short_name
and passed as a scenario, while provided action is linters.

We can't use this check for openstack-ansible repo itself, since we have
a lot of linters debt, so we need to fix things inside roles first.

Depends-On: https://review.opendev.org/679238
Depends-On: https://review.opendev.org/679105
Change-Id: I24880ab92f89ca02f6b09a63da9f03dd5b230083