Install galera client for placement migration
The placement migration tool interacts directly with the galera database
and needs the client libraries and command line tools in order to do this.
On a metal deploy the placement and database may be co-located so these
libraries are present, but on an LXC deployment they are not available
in the placement container.

During upgrade from stein to train received the error:
"Failed to connect to NOVA_API database", "/openstack/venvs/placement
-20.0.0.0rc2.dev17/lib/python3.6/site-packages/placement_db_tools/
mysql-migrate-db.sh: line 79: mysql: command not found


Change-Id: I85b97d1f503bfbda5a893b9afbf74bd28f732882