Install appropriate tempest releases on stable branches
To this point OSA has used the most recent tempest release (which is
itself branchless) during CI tests of OSA stable branches.

This is no longer possible becasue contemporary releases of tempest are
dropping py2 compatibility and have py3-only dependencies and therefore
cannot be installed on OSA stable branches prior to Train which have no
py3 support.

This patch includes the file repo_packages/openstack_services.yml
into the os-tempest-install playbook which has the effect of defining
requirements_git_install_branch, which in turn means that
tempest_upper_constraints_url no longer defaults to the master branch of
u-c. This means that a stable-branch-appropriate set of u-c is applied
when installing the python packages for tempest, effectively holding tempest
back at a release and subsequent set of dependancies that should work
in the context of the stable branch.

This changes should not affect the behaviour on OSA master branch
where tempest is installed using source code from git. This is unchanged
due to the presence of variables in repo_packages/openstack_testing.yml
on OSA master which override the defaults in the os_tempest ansible role.
These variables are removed on stable branches as part of the OSA release
process.

Change-Id: Ic3aaae97de86876a09377781a316f2468fcaa2b2