Ensure utility binaries link to the latest versions
The previous code would not move the openstack cli symlinks from
old venv to new ones at either minor or major release upgrades.

This patch changes the code to use a native ansible module instead
of shell commands. The symlinks are now moved to the current venv
and the tasks are idempotent.

Change-Id: If10f83814d1934b430749a33efdbf20408e2fa3d