Use zuul cached repos for openstack services
In CI there are cached repos for the openstack services as well as
the OSA repos. We are currently using the cached repos for OSA ansible
roles, but not for the wider openstack services which are installed
from specific git repository SHA as defined by
playbooks/defaults/repo_packages/openstack_services.yml.

This patch parses the openstack_services.yml file and generates a
set of OSA override variables to use the locally cached git repos
instead of those at opendev.org. Network traffic is reduced and the
impact of OSA jobs on the opendev git servers is eased which has
caused problems when the proposal bot makes changes to all OSA repos
simultaneously.

Depends-On: https://review.opendev.org/712459
Change-Id: I88087e3b4d6a688670e40165f127049956705801