ceph: run handlers after common
When you're running ceph-install with upgrade_ceph_packages, the
packages get upgraded in the ceph-common, but some services aren't
getting restarted to account for the new version.

According to ceph-ansible/site.yml.sample, the order is ceph-common
applied to all hosts, then later in the "hosts: mons" or "hosts: osd"
the handlers get applied, after ceph-facts and before ceph-config.

Change-Id: I3e95b7fad6eda9f689287efe16c35418a9e7e561