Add integrated test for Ubuntu using ML2/OVN driver
This patchset implements an integrated test for a new aio_metal_ovn
scenario, which deploys the ML2/OVN driver and associated services. The
basic server ops tempest test is also included. Expectations are that
the deployed cirros VM will be reachable behind a floating IP internally
for testing purposes.

add testing for both bionic and focal

Depends-On: https://review.opendev.org/#/c/750802

Change-Id: I1ff22f8b42bfb421a7af3ee0d6ffc3c25e7d4d91